requirements.txt
fastapi 
uvicorn 
supabase
pyjwt
psycopg[binary]
requests
dotenv
langchain
langchain-groq
langgraph
langchain-core
structlog
pydantic_settings
git-filter-repo
streamlit
watchdog
psycopg2
prometheus-client